refactor: 기존의 Hibernate를 이용한 Repository 코드 삭제
