+++ backend/src/main/java/com/woowacourse/auth/service/AuthService.java
@@ -6,8 +6,7 @@ import com.woowacourse.auth.exception.AuthorizationFailureException;
 import com.woowacourse.auth.support.JwtTokenProvider;
 import com.woowacourse.moragora.entity.user.RawPassword;
 import com.woowacourse.moragora.entity.user.User;
-import com.woowacourse.moragora.repository.user.UserRepository;
-import org.springframework.beans.factory.annotation.Qualifier;
+import com.woowacourse.moragora.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,7 +18,7 @@ public class AuthService {
     private final UserRepository userRepository;
 
     public AuthService(final JwtTokenProvider jwtTokenProvider,
-                       @Qualifier("userSpringJpaRepository") final UserRepository userRepository) {
+                       final UserRepository userRepository) {
         this.jwtTokenProvider = jwtTokenProvider;
         this.userRepository = userRepository;
     }
