+++ backend/src/main/java/com/woowacourse/moragora/repository/AttendanceRepository.java
@@ -1,13 +1,12 @@
-package com.woowacourse.moragora.repository.attendance;
+package com.woowacourse.moragora.repository;
 
 import com.woowacourse.moragora.entity.Attendance;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttendanceRepository {
-
-    Attendance save(final Attendance attendance);
+public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
 
     List<Attendance> findByParticipantId(final Long participantId);
