+++ backend/src/main/java/com/woowacourse/moragora/repository/MeetingRepository.java
@@ -1,0 +1,7 @@
+package com.woowacourse.moragora.repository;
+
+import com.woowacourse.moragora.entity.Meeting;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MeetingRepository extends JpaRepository<Meeting, Long> {
+}
