+++ backend/src/main/java/com/woowacourse/moragora/repository/ParticipantRepository.java
@@ -1,11 +1,11 @@
-package com.woowacourse.moragora.repository.participant;
+package com.woowacourse.moragora.repository;
 
 import com.woowacourse.moragora.entity.Participant;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParticipantSpringJpaRepository extends JpaRepository<Participant, Long>, ParticipantRepository {
+public interface ParticipantRepository extends JpaRepository<Participant, Long> {
 
     List<Participant> findByMeetingId(final Long meetingId);
