--- backend/src/main/java/com/woowacourse/moragora/repository/meeting/MeetingRepository.java
@@ -1,11 +1,0 @@
-package com.woowacourse.moragora.repository.meeting;
-
-import com.woowacourse.moragora.entity.Meeting;
-import java.util.Optional;
-
-public interface MeetingRepository {
-
-    Meeting save(final Meeting meeting);
-
-    Optional<Meeting> findById(final Long id);
-}
