+++ backend/src/test/java/com/woowacourse/moragora/repository/AttendanceRepositoryTest.java
@@ -3,7 +3,6 @@ package com.woowacourse.moragora.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.woowacourse.moragora.entity.Attendance;
-import com.woowacourse.moragora.repository.attendance.AttendanceHibernateRepository;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -15,10 +14,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
 @Transactional
-class AttendanceHibernateRepositoryTest {
+class AttendanceRepositoryTest {
 
     @Autowired
-    private AttendanceHibernateRepository attendanceRepository;
+    private AttendanceRepository attendanceRepository;
 
     @DisplayName("미팅 참가자의 누적 출석정보를 조회한다")
     @Test
