+++ backend/src/test/java/com/woowacourse/moragora/repository/MeetingRepositoryTest.java
@@ -3,7 +3,6 @@ package com.woowacourse.moragora.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.woowacourse.moragora.entity.Meeting;
-import com.woowacourse.moragora.repository.meeting.MeetingHibernateRepository;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
@@ -14,10 +13,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
 @Transactional
-class MeetingHibernateRepositoryTest {
+class MeetingRepositoryTest {
 
     @Autowired
-    private MeetingHibernateRepository meetingRepository;
+    private MeetingRepository meetingRepository;
 
     @DisplayName("미팅 방을 저장한다.")
     @Test
